refactor: add InputLabel folder and update dep
