+++ src/ui/components/InputLabel/styledInputLabel.ts
